+++ spec/models/jump_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+describe Jump do
+  it_should_behave_like 'a basic model', 'Jump'
+end