+++ CHANGELOG.md
@@ -4,6 +4,17 @@ for changelogs and semantic versioning available at:
 * https://keepachangelog.com/en/0.3.0/
 * https://semver.org/
 
+## [0.1.4] - 2018-07-12, 2018-07-23, 2018-07-24
+### Added
+- Character specs
+- Jump table, model, factory, specs
+- Jump spec behavior factored out to helper
+- Character "
+- CharacterType and JumpTypes updated to allow accessing data
+
+### Todo (Completed)
+- Fix Char spec to match updated jump specs
+
 ## [0.1.3] - 2018-07-06
 ### Added
 - Character primitive attributes